+++ include/frepple/tags.h
@@ -106,6 +106,7 @@ class Tags {
   static const Keyword headerstart;
   static const Keyword hidden;
   static const Keyword id;
+  static const Keyword individualPoolResources;
   static const Keyword interruption;
   static const Keyword interruptions;
   static const Keyword ip_flag;