+++ src/tags.cpp
@@ -105,6 +105,7 @@ const Keyword Tags::headeratts("headeratts");
 const Keyword Tags::headerstart("headerstart");
 const Keyword Tags::hidden("hidden");
 const Keyword Tags::id("id");
+const Keyword Tags::individualPoolResources("individualPoolResources");
 const Keyword Tags::interruption("interruption");
 const Keyword Tags::interruptions("interruptions");
 const Keyword Tags::ip_flag("ip_flag");